feat: Create PtcRunner.Json public API and deprecate PtcRunner
Implement Phase 1.1/1.4 of the API refactor plan per docs/api-refactor-plan.md:

- Create new PtcRunner.Json module as the public API for JSON DSL programs
- Move full implementation logic (run/2, run!/2, format_error/1) from
  PtcRunner to PtcRunner.Json
- Update PtcRunner module to use deprecation delegates pointing to Json module
- Update all test files to use PtcRunner.Json.run instead of deprecated API

The existing PtcRunner.run/2, PtcRunner.run!/2, and PtcRunner.format_error/1
now delegate to their Json counterparts with @deprecated annotations, ensuring
backward compatibility while guiding users toward the new namespaced API.

All 549 tests pass. This enables future multi-language support (e.g. PTC-Lisp).

Fixes #102
