+++ test/ptc_runner/e2e_test.exs
@@ -20,7 +20,7 @@ defmodule PtcRunner.E2ETest do
         ]
       }
 
-      assert {:ok, result, _metrics} = PtcRunner.run(program_json, context: context)
+      assert {:ok, result, _metrics} = PtcRunner.Json.run(program_json, context: context)
 
       assert is_list(result)
       assert length(result) == 2
@@ -39,7 +39,7 @@ defmodule PtcRunner.E2ETest do
         ]
       }
 
-      assert {:ok, result, _metrics} = PtcRunner.run(program_json, context: context)
+      assert {:ok, result, _metrics} = PtcRunner.Json.run(program_json, context: context)
       assert result == 60
     end
 
@@ -55,7 +55,7 @@ defmodule PtcRunner.E2ETest do
         ]
       }
 
-      assert {:ok, result, _metrics} = PtcRunner.run(program_json, context: context)
+      assert {:ok, result, _metrics} = PtcRunner.Json.run(program_json, context: context)
       assert result == 2
     end
 
@@ -74,7 +74,7 @@ defmodule PtcRunner.E2ETest do
         ]
       }
 
-      assert {:ok, result, _metrics} = PtcRunner.run(program_json, context: context)
+      assert {:ok, result, _metrics} = PtcRunner.Json.run(program_json, context: context)
       assert result["name"] == "Bob"
       assert result["years_employed"] == 7
     end
@@ -92,7 +92,7 @@ defmodule PtcRunner.E2ETest do
         ]
       }
 
-      assert {:ok, result, _metrics} = PtcRunner.run(program_json, context: context)
+      assert {:ok, result, _metrics} = PtcRunner.Json.run(program_json, context: context)
       assert result["name"] == "Book"
       assert result["price"] == 15
     end
@@ -110,7 +110,7 @@ defmodule PtcRunner.E2ETest do
         ]
       }
 
-      assert {:ok, result, _metrics} = PtcRunner.run(program_json, context: context)
+      assert {:ok, result, _metrics} = PtcRunner.Json.run(program_json, context: context)
       assert result == ["Apple", "Book", "Laptop"]
     end
 
@@ -127,7 +127,7 @@ defmodule PtcRunner.E2ETest do
         ]
       }
 
-      assert {:ok, result, _metrics} = PtcRunner.run(program_json, context: context)
+      assert {:ok, result, _metrics} = PtcRunner.Json.run(program_json, context: context)
       assert is_list(result)
       assert length(result) == 3
       prices = Enum.map(result, & &1["price"])
@@ -150,7 +150,7 @@ defmodule PtcRunner.E2ETest do
         ]
       }
 
-      assert {:ok, result, _metrics} = PtcRunner.run(program_json, context: context)
+      assert {:ok, result, _metrics} = PtcRunner.Json.run(program_json, context: context)
 
       # Constraint assertions - not exact values
       assert is_list(result)
@@ -171,7 +171,7 @@ defmodule PtcRunner.E2ETest do
         ]
       }
 
-      assert {:ok, result, _metrics} = PtcRunner.run(program_json, context: context)
+      assert {:ok, result, _metrics} = PtcRunner.Json.run(program_json, context: context)
       assert result == 60
     end
 
@@ -188,7 +188,7 @@ defmodule PtcRunner.E2ETest do
         ]
       }
 
-      assert {:ok, result, _metrics} = PtcRunner.run(program_json, context: context)
+      assert {:ok, result, _metrics} = PtcRunner.Json.run(program_json, context: context)
       assert result == 2
     end
   end
